all seeder add

+++ database/migrations/2022_04_18_063439_create_product_materials_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('product_id');
             $table->unsignedBigInteger('material_id');
-            $table->integer('quantity')->default(0);
+            $table->float('quantity')->default(0);
             $table->timestamps();
             $table->foreign("product_id")->references("id")->on("products");
             $table->foreign("material_id")->references("id")->on("materials");
